+++ Sources/TidalAPIInterchangeKit/API/Models/TidalArtwork.swift
@@ -29,3 +29,44 @@ public struct TidalArtwork: Codable, Equatable, Hashable, Sendable {
         self.type = type
     }
 }
+
+public extension TidalArtwork {
+
+    var smallestImage: TidalArtworkFile? {
+        guard let attributes else { return nil }
+
+        return attributes.files.min { lhsFile, rhsFile in
+            (lhsFile.meta?.width ?? 0) * (lhsFile.meta?.height ?? 0) < (rhsFile.meta?.width ?? 0) * (rhsFile.meta?.height ?? 0)
+        }
+    }
+
+    var largestImage: TidalArtworkFile? {
+        guard let attributes else { return nil }
+
+        return attributes.files.max { lhsFile, rhsFile in
+            (lhsFile.meta?.width ?? 0) * (lhsFile.meta?.height ?? 0) < (rhsFile.meta?.width ?? 0) * (rhsFile.meta?.height ?? 0)
+        }
+    }
+
+    func smallestImageWithSizeAtLeast(width: Int, height: Int) -> TidalArtworkFile? {
+        guard let attributes else { return nil }
+
+        return attributes.files.reduce(nil) {
+            guard let fileWidth = $1.meta?.width,
+                  let fileHeight = $1.meta?.height,
+                  fileWidth >= width,
+                  fileHeight >= height else { return $0 }
+
+            guard let accumulatedFile = $0,
+                  let accumulatedWidth = accumulatedFile.meta?.width,
+                  let accumulatedHeight = accumulatedFile.meta?.height else { return $1 }
+
+            if fileWidth < accumulatedWidth,
+               fileHeight < accumulatedHeight {
+                return $1
+            } else {
+                return accumulatedFile
+            }
+        }
+    }
+}